+++ Source/GOATracer/GOATracer.sln
@@ -1,6 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cs-raytracer", "cs-raytracer.csproj", "{0FB92275-89FF-4B17-A44E-CFA53299781F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GOATracer", "GOATracer.csproj", "{0FB92275-89FF-4B17-A44E-CFA53299781F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
