+++ Source/GOATracer/MainWindow.axaml
@@ -3,7 +3,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="cs_raytracer.MainWindow"
-        Title="cs_raytracer">
+        x:Class="GOATracer.MainWindow"
+        Title="GOATracer">
     Welcome to Avalonia!
 </Window>
